Refactor request body handling by consolidating `RequestBody` and `FormData` into a unified structure. Removed the `request_body.rs` file and updated imports across the codebase to reflect the new organization. Enhanced `FormData` functionality and adjusted related tests for consistency.
